Create ZTH.yaml
The Zethereum - ZTH cryptocurrency is a token issued on the Signum blockchain in December 2021. The Signum blockchain allows you to issue smart tokens that can interact with smart contracts. Now ZTH is its own token in the Signum blockchain with the asset ID: 9518219425200752102, and it is controlled by the GrandPoolZTH and GrandRewardZTH smart contracts.

The ZTH token was issued in the amount of 100,000,000 units. 70% of the minted coins are available for mining, another 20% for placing bets, and 10% are aimed at the development of the project. In automatic mode, every 6 hours, smart contracts distribute 1000 ZTH between all mining and staking participants.